Create Get Locked AD User Accounts and Unlock
Retrieves all AD accounts that have account lockout times (are currently locked out). The ForEach enumerates all of the locked accounts and unlocks them with the Unlock-ADAccount cmdlet. A certain set of the locked accounts can be unlocked by adding the appropriate property to be filtered in the Where-Object (?) statement.