Create Get Disk Volumes From Remote AD Computers Using PS
Creates a remote connections/sessions with all AD computers other than the local host and runs a command against each of the sessions to retrieve disk volume information. The persistent session is then manually discarded. The disk volume information in the variable, $drives, is then formatted and output as a table. For fun, the disk volume information is enumerated to output certain phrases depending on how much space is left in each disk volume.

One of the more interesting things I learned was that you must first Sort by a property in order to Group by that property; otherwise, "GroupBy" will not work properly.