Create Get Data From Remote AD Computers Using PS-Remoting
Added additional functions to retrieve and output different types of data (only if run separately). There does not seem to be a way to output data from different commands in one PS script. PS may be treating each output as entering data into a table and then showing it to the user. In this case, it makes sense that only one output would be allowed in order to prevent conflicting column names in the output / table.